backend/requirements.txt
# Web框架和API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
fastapi-limiter>=0.1.6

loom-agent==0.0.4
tiktoken

# 数据库和ORM
SQLAlchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
sqlparse>=0.4.4

# Redis和缓存
redis>=5.2.0

# 认证和安全
PyJWT>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
cryptography>=44.0.0

# 文件处理
python-multipart>=0.0.12
python-docx>=1.1.2

# 对象存储
minio>=7.2.10

# 任务队列和调度
celery[redis]>=5.4.0
apscheduler>=3.11.0
cron-validator>=1.0.8
croniter>=1.4.1
nest-asyncio>=1.6.0

# 数据处理
pandas>=2.2.3
numpy>=2.1.0

# HTTP客户端和网络
requests>=2.32.0
httpx>=0.24.0
aiohttp>=3.8.0
aiofiles>=23.1.0

# 日志和监控
structlog>=24.4.0

# 可视化
matplotlib>=3.10.0
seaborn>=0.13.2

# 工具库
python-dotenv>=1.0.1
unidecode>=1.3.8
pytz>=2023.3
aiosqlite>=0.19.0

# 数据库连接器
PyMySQL>=1.1.1

# 系统和进程监控
psutil>=5.9.0
